requirements.txt
flask==2.3.3
requests==2.31.0
oauth1==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
pycryptodome==3.19.0